% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_vector.R
\name{mi_vector}
\alias{mi_vector}
\title{Calculate mutual information between vectors}
\usage{
mi_vector(x, y, mitype = "kde", minorm = F, autoswitch = F, lc = T)
}
\description{
Calculate mutual information between vectors
}
\details{
Results are expressed in nats.
}
