% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics.R
\name{mics}
\alias{mics}
\title{Cross-spectrum based on mutual information}
\usage{
mics(spectro1 = NULL, spectro2 = NULL, feature1 = NULL,
  feature2 = NULL, plot = F, zerofill = T, lagw = NULL, ...)
}
\description{
Cross-spectrum based on mutual information
}
