% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_spectro.R
\name{signal_spectro}
\alias{signal_spectro}
\title{Short-term FFT only for signal portions of a sound}
\usage{
signal_spectro(sound, plot = F, segment_params = list(wl = 512, ovlp =
  0, threshold = 5), ...)
}
\description{
Short-term FFT only for signal portions of a sound
}
