% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MI.R
\name{sim_MI}
\alias{sim_MI}
\title{Compare two sounds based using mutual information}
\usage{
sim_MI(sound1 = NULL, sound2 = NULL, features = NULL, miN = 10,
  d1 = NULL, d2 = NULL, type = "jackknife", nlcc = F, plot = F,
  ...)
}
\arguments{
\item{sound1}{Sound 1. Wave object.}

\item{sound2}{Sound 2. Wave object.}

\item{features}{Features data frame.}

\item{miN}{Number of neighbors for estimation of mutual information}

\item{d1}{Discretized spectrogram of sound 1.}

\item{d2}{Discretized spectrogram of sound 2.}

\item{type}{Type of mutual information estimator. See details.}

\item{nlcc}{Should the non-linear correlation coefficient be reported?. See details.}

\item{plot}{Plot}

\item{...}{Additional arguments passed to infotheo::mutinformation.}

\item{norm}{Normalize mutual information?}

\item{align}{Should time traces be aligned?}
}
\description{
Compare two sounds based using mutual information
}
\details{
The types of estimation of mutual information are:

symba -  Estimate mutual information using seewave::symba.
infotheo - Estimate mutual information using package infotheo.

The non-linear correlation coefficient is that of XXXX.
}
