% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_batch.R
\name{similarity_batch}
\alias{similarity_batch}
\title{Take similarity measurements in batch}
\usage{
similarity_batch(sound1, sound2, s1.sylls, s2.sylls,
  spectroargs = list(wl = 512, ovlp = 0, wn = "hanning"),
  alignment = "lc", datatype = "spectro")
}
\description{
Take similarity measurements in batch
}
