% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_features.R
\name{spectral_features}
\alias{spectral_features}
\title{Get spectral features of a wave file}
\usage{
spectral_features(wave, windowargs = list(wl = 512, ovlp = 0, threshold =
  5), ...)
}
\arguments{
\item{wave}{Wave object.}

\item{...}{Other \link[misound]{signal_spectro} arguments.}

\item{thresh}{Threshold for separating signal from noise. % of maximum
amplitude.}
}
\description{
Get spectral features of a wave file
}
